Rewrite Modal
